requirements.txt
flask==3.1.1
opencv-contrib-python-headless==4.11.0.86
numpy==2.2.6
Pillow==11.2.1
